refactor(eigen-py): renamed eigen-py package to eigen-client due to naming constraints with pypi
