+++ libs/eigen-py/README.md
@@ -1,12 +1,22 @@
 # EigenDB's Official Python API 🐍
 
-A Python wrapper for EigenDB's REST API. 
+![](https://img.shields.io/badge/Python-3776AB?style=flat&logo=python&logoColor=white) [![](https://img.shields.io/pypi/v/eigen-client)](https://pypi.org/project/eigen-client/)
+
+![OpenAI](https://img.shields.io/badge/OpenAI-fff?style=for-the-badge&logo=openai&logoColor=black) [![Ollama](https://img.shields.io/badge/Ollama-fff?logo=ollama&style=for-the-badge&logoColor=000)](#)
+
+A Python client for EigenDB's REST API. 
+
+### Installation
+
+```bash
+pip install eigen-client
+```
 
 ### Example usage:
 ```py
 import os
-from eigen_py.index import Index
-from eigen_py.data_types import Document
+from eigen_client.index import Index
+from eigen_client.data_types import Document
 
 index = Index(
     url="http://localhost:8080",
