+++ libs/eigen-py/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
-name = "eigen-py"
-version = "1.0.0"
+name = "eigen-client"
+version = "1.0.2"
 authors = [
   { name="Ryan Awad", email="ryanawad@yahoo.com" },
 ]
@@ -21,5 +21,5 @@ dependencies = [
 ]
 
 [project.urls]
-Homepage = "https://github.com/Eigen-DB/eigen-db/libs/eigen-py"
+Homepage = "https://github.com/Eigen-DB/eigen-db/tree/main/libs/eigen-py"
 Issues = "https://github.com/Eigen-DB/eigen-db/issues"
