+++ libs/eigen-py/src/eigen_client/supported_models.py
